Code Update and Restructuring
