+++ src/ansys/fluent/core/rpvars.py
@@ -6,8 +6,8 @@ interfaces: solver settings objects and task-based meshing workflow.
 """
 from typing import Any, List
 
-from ansys.fluent.core.allowed_name_error_msg import allowed_name_error_message
 import ansys.fluent.core.filereader.lispy as lispy
+from ansys.fluent.core.solver.allowed_name_error_msg import allowed_name_error_message
 
 
 class RPVars:
